Don't do equality check of secret key in simultaneous verification of multiple disclosure proofs
When disclosing attributes coming from a credential that does and a credential that doesn't
involve a keyshare server, the secret key (zero'th attribute) of the credentials do not
in fact coincide, so we cannot expect a proof of equality to hold.
When verifying proofs involving either none or one keyshare server the secret keys and
their proofs of knowledge should coincide, but we do not know which credentials involve
which keyshare servers, so we cannot verify that the proofs indeed show that the credentials
share their secret keys. Therefore, we do not do this here anymore: instead, this is now
the responsibility of the caller (i.e., irmago).