feat: add support for contextual linkability (or 1-anonymity)
This adds a feature where verifiers can require clients to include extra integers in their disclosure proofs, which are stable per (preconfigured) epoch and verifier name, but distinct otherwise. These are computed as R = H(epoch || verifier || 0)^attr, where H is a hash function and attr is one of the user's undisclosed attributes. The user includes a proof that R is of this form, while keeping attr undisclosed. Thus if the user does a second disclosure during the same epoch to the same verifier, she becomes linkable because she has to send the same R as the previous time - that is, this implements 1-anonymity. However, when disclosing instead a second time during another epoch or to another verifier, R is different, so in that case unlinkability is preserverd.

The used attribute should be issued using randomblind issuance, so that it is large, guaranteed to be distinct from all other issued attributes, and unknown to the issuer.

The 0 in the hash is meant for future compatibility with a possible extension of this from 1-anonymity to k-anonymity.