Modified core of gabi to provide functionality for range proofs.
I have chosen to separate the range proof outside of gabis normal
prooflist structure, as it requires significant description of what
should be proven to be given by the verifier, which just does not fit in the
prooflist model used by gabi for disclosure and issuance proofs (which
are essentially self-describing). A more full integration would require
significant restructuring of how the core of gabi thinks about proofs
and verifying them.