test: increase group sizes used in some keyproof tests to avoid degenerate panics
Previously, some tests would accidentally factor the modulus [0] or hit 0 in Z/pZ [1] due to the miniscule group sizes (~100 elements or less) used in the tests, triggering a panic which would cause the test to fail. This commit generates once a set of parameters that is used across the tests to instantiate groups that are still small for speed, but large enough that they should avoid with high probability these degenerate cases.

[0]: e.g. https://github.com/privacybydesign/gabi/blob/6113e0d3ec8106175ab705187813a37604309c2e/keyproof/squarefree.go#L27
[1]: https://github.com/privacybydesign/gabi/blob/6113e0d3ec8106175ab705187813a37604309c2e/keyproof/primeproof.go#L182