feat: add functions to be used by irmaclient to compute keyshare protocol messages
This adds the functions KeyshareUserCommitmentRequest() and KeyshareUserResponseRequest(), which can be used by the irmaclient to compute its first and second message respectively to the keyshare server in the new keyshare protocol. These functions abstract the cryptographic details of the protocol away from the irmaclient, and accept parameters that the irmaclient will mostly already have. They contain code that was previously in the testNewKeyshareResponse() test helper function, which has therefore become shorter and closer to the actual future implementation in the irmaclient.